refactor(admin): split live repos panels
Why: keep the live code route entrypoint focused on stream wiring while separating repo sorting and panel rendering into named modules.

Verify: pnpm exec prettier --check apps/admin/src/app/code/live-repos.tsx apps/admin/src/app/code/live-repos-model.ts apps/admin/src/app/code/live-repos-sections.tsx; git diff --check -- apps/admin/src/app/code/live-repos.tsx apps/admin/src/app/code/live-repos-model.ts apps/admin/src/app/code/live-repos-sections.tsx; pnpm --filter @anipotts/admin typecheck; pnpm --filter @anipotts/admin lint; pnpm --filter @anipotts/admin build; pnpm validate
